[ENH]: update blockfile interfaces to enable new sorted insert mode
Also re-org'ed some of the type files.
