+++ rust/blockstore/src/arrow/concurrency_test.rs
@@ -1,6 +1,9 @@
 #[cfg(test)]
 mod tests {
-    use crate::arrow::{config::TEST_MAX_BLOCK_SIZE_BYTES, provider::ArrowBlockfileProvider};
+    use crate::{
+        arrow::{config::TEST_MAX_BLOCK_SIZE_BYTES, provider::ArrowBlockfileProvider},
+        BlockfileWriterOptions,
+    };
     use chroma_cache::new_cache_for_test;
     use chroma_storage::{local::LocalStorage, Storage};
     use rand::Rng;
@@ -22,7 +25,10 @@ mod tests {
                     block_cache,
                     sparse_index_cache,
                 );
-                let writer = blockfile_provider.create::<&str, u32>().unwrap();
+                let writer = future::block_on(
+                    blockfile_provider.get_writer::<&str, u32>(BlockfileWriterOptions::default()),
+                )
+                .unwrap();
                 let id = writer.id();
                 // Generate N datapoints and then have T threads write them to the blockfile
                 let range_min = 10;
