+++ rust/index/src/fulltext/types.rs
@@ -547,11 +547,11 @@ mod tests {
     use chroma_blockstore::provider::BlockfileProvider;
     use tantivy::tokenizer::NgramTokenizer;
 
-    #[test]
-    fn test_new_writer() {
+    #[tokio::test]
+    async fn test_new_writer() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let tokenizer = Box::new(TantivyChromaTokenizer::new(
             NgramTokenizer::new(1, 1, false).unwrap(),
         ));
@@ -562,8 +562,8 @@ mod tests {
     #[tokio::test]
     async fn test_new_writer_then_reader() {
         let provider = BlockfileProvider::new_memory();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
         let freq_blockfile_id = freq_blockfile_writer.id();
         let pl_blockfile_id = pl_blockfile_writer.id();
 
@@ -590,8 +590,8 @@ mod tests {
     #[tokio::test]
     async fn test_index_and_search_single_document() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
 
@@ -629,8 +629,8 @@ mod tests {
     #[tokio::test]
     async fn test_repeating_character_in_query() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
 
@@ -662,8 +662,8 @@ mod tests {
     #[tokio::test]
     async fn test_query_of_repeating_character() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
 
@@ -696,8 +696,8 @@ mod tests {
     #[tokio::test]
     async fn test_repeating_character_in_document() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
 
@@ -729,8 +729,8 @@ mod tests {
     #[tokio::test]
     async fn test_search_absent_token() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
 
@@ -762,8 +762,8 @@ mod tests {
     #[tokio::test]
     async fn test_multiple_candidates_within_document() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
 
@@ -802,8 +802,8 @@ mod tests {
     #[tokio::test]
     async fn test_multiple_simple_documents() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
 
@@ -839,8 +839,8 @@ mod tests {
     #[tokio::test]
     async fn test_multiple_complex_documents() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
 
@@ -884,8 +884,8 @@ mod tests {
     #[tokio::test]
     async fn test_index_multiple_character_repeating() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
 
@@ -930,8 +930,8 @@ mod tests {
     #[tokio::test]
     async fn test_index_special_characters() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
 
@@ -974,8 +974,8 @@ mod tests {
     #[tokio::test]
     async fn test_get_frequencies_for_token() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
 
@@ -1017,8 +1017,8 @@ mod tests {
     #[tokio::test]
     async fn test_get_all_results_for_token() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
 
@@ -1060,8 +1060,8 @@ mod tests {
     #[tokio::test]
     async fn test_update_document() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
 
@@ -1104,8 +1104,8 @@ mod tests {
     #[tokio::test]
     async fn test_delete_document() {
         let provider = BlockfileProvider::new_memory();
-        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().unwrap();
-        let freq_blockfile_writer = provider.create::<u32, String>().unwrap();
+        let pl_blockfile_writer = provider.create::<u32, Vec<u32>>().await.unwrap();
+        let freq_blockfile_writer = provider.create::<u32, String>().await.unwrap();
         let pl_blockfile_id = pl_blockfile_writer.id();
         let freq_blockfile_id = freq_blockfile_writer.id();
